more structural refactoring
